+++ squads/forms.py
@@ -22,7 +22,7 @@ class CustomClearableFileInput(ClearableFileInput):
 class SquadsGroupForm(forms.ModelForm):
     class Meta:
         model = Groups
-        fields = ["name", "description", "require_approval", "is_active", "image"]
+        fields = ["name", "description", "req_approve", "is_active", "image"]
         widgets = {
             "image": CustomClearableFileInput(),
         }
@@ -35,9 +35,7 @@ class SquadsGroupForm(forms.ModelForm):
         self.fields["description"].widget.attrs.update(
             {"class": "form-control", "maxlength": "1000", "rows": "5"}
         )
-        self.fields["require_approval"].widget.attrs.update(
-            {"class": "form-check-input"}
-        )
+        self.fields["req_approve"].widget.attrs.update({"class": "form-check-input"})
         self.fields["is_active"].widget.attrs.update({"class": "form-check-input"})
 
     def clean_name(self):