+++ squads/models/__init__.py
@@ -1,4 +1,4 @@
 from .filters import AssetsFilter, SkillSetFilter, SquadFilter, SquadGroup
 from .general import General
-from .groups import Groups, Pending
-from .memberships import Memberships
+from .groups import Groups
+from .member import Memberships, Pending