+++ squads/views/groups.py
@@ -11,8 +11,8 @@ from squads.app_settings import SQUADS_EMPTY_IMAGE
 from squads.forms import CommentForm, SquadsGroupForm
 from squads.hooks import get_extension_logger
 from squads.models.filters import SquadGroup
-from squads.models.groups import Groups, Pending
-from squads.models.memberships import Memberships
+from squads.models.groups import Groups
+from squads.models.member import Memberships, Pending
 from squads.views._core import add_info_to_context
 
 logger = get_extension_logger(__name__)